requirements.txt
InquirerPy==0.3.4
pandas==2.2.2
pyfiglet==1.0.2
PyYAML==6.0.2
Requests==2.32.3
setuptools==70.3.0
typer==0.12.5
wasabi==1.1.3
